mobile-api/requirements.txt
# Mobile API dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

# Voice Pipeline
openai>=1.10.0
google-cloud-texttospeech>=2.16.0

# WebSocket
websockets>=12.0

# Shared with soul-kun (inherited via sys.path)
# pg8000, flask, etc.
